requirements.txt
###### Requirements without Version Specifiers ######
unittest2==1.1.0
requests2==2.16.0
nose2==0.9.2
